add hypervalent geom hint
sulfone/phosphine oxide "hints" for cdxml parsing.

This requires reading in a dictionary of atom id -> atom text to check which atoms need a wedge/hash. For a phosphine oxide, it is the P-O, for a sulfone, only one of the S-O bonds.

It is a little clunky, but it does seem to work.